+++ familyhelper-ui-view/packages/webapp-mobile/src/views/nodes/comvisual/CompregFallback.vue
@@ -1,14 +1,19 @@
 <template>
   <div class="compreg-fallback-container">
-    <div class="main-container">
-      <div class="placeholder">
-        当你看到这个页面的时候，意味着 router.component.key 没有正确配置
+    <border-layout-panel class="border-layout-panel">
+      <div class="main-container">
+        <placeholder-panel
+          text="当你看到这个页面的时候，意味着 router.component.key 没有正确配置"
+        />
       </div>
-    </div>
+    </border-layout-panel>
   </div>
 </template>
 
 <script setup lang="ts">
+import PlaceholderPanel from '@/components/comvisual/layout/placeholderPanel/PlaceholderPanel.vue'
+import BorderLayoutPanel from '@/components/elementPlus/layout/borderLayoutPanel/BorderLayoutPanel.vue'
+
 defineOptions({
   name: 'CompregFallbackComponent',
 })
@@ -23,18 +28,5 @@ defineOptions({
 .main-container {
   width: 100%;
   height: 100%;
-}
-
-.placeholder {
-  width: 100%;
-  height: 100%;
-  text-align: center;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  font-size: 24px;
-  font-weight: bold;
-  color: #bfbfbf;
-  user-select: none;
 }
 </style>