+++ familyhelper-ui-view/packages/webapp-mobile/src/views/nodes/comvisual/welcome/Welcome.vue
@@ -1,10 +1,17 @@
 <template>
   <div class="welcome-container">
-    <div class="placeholder">欢迎使用 VIM 框架</div>
+    <border-layout-panel class="border-layout-panel">
+      <div class="main-container">
+        <placeholder-panel text="欢迎使用 VIM 框架" />
+      </div>
+    </border-layout-panel>
   </div>
 </template>
 
 <script setup lang="ts">
+import PlaceholderPanel from '@/components/comvisual/layout/placeholderPanel/PlaceholderPanel.vue'
+import BorderLayoutPanel from '@/components/comvisual/layout/borderLayoutPanel/BorderLayoutPanel.vue'
+
 defineOptions({
   name: 'WelcomeComponent',
 })
@@ -16,16 +23,8 @@ defineOptions({
   height: 100%;
 }
 
-.placeholder {
+.main-container {
   width: 100%;
   height: 100%;
-  text-align: center;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  font-size: 24px;
-  font-weight: bold;
-  color: #bfbfbf;
-  user-select: none;
 }
 </style>